ResponsibleAI - Input Moderation - Renaming of files for better clarity and alignment to functionality
