+++ packages/components/nodes/moderation/OpenAIModeration/openai-moderation.png
