+++ packages/components/nodes/moderation/SimplePromptModeration/simple_moderation.png
